LayoutTests/js/parse-error-external-script-in-eval-expected.txt

CONSOLE MESSAGE: line 2: SyntaxError: Expected an identifier but found 'error' instead
This tests that a line number and error message is printed for a parse error in a separate source file using eval().
